core: add support to track whether soft device is enabled
When a soft device is present and enabled some resources are restricted.
The soft device API includes a function that can be used to determine
whether it's enabled, but the linkage of that function prevents
providing a weak definition for applications that don't use the SoC
library.  Provide API in systemState to track this independently.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>