core: add API to extract die temperature
Add API to get the die temperature in standard cCel, defaulting to the
TEMP peripheral but delegating to the soft-device SoC library API when
the TEMP peripheral is restricted.

This commit also demonstrates how to provide SD overrides for restricted
resources.  The subtle bit is that the library providing the non-weak
implementation (libnrfsd.a) must appear before the library that uses the
non-weak implementation (libnrfcxx.a) , else the uses will be bound to
the weak definition which is incorrect.  Further, the weak definition
must be explicitly undefined so that the weak definition found
internally to libnrfcxx.a is not preferred.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>