sensor/adc: add a calibration state machine
It's annoying to have to implement calibration support in every
ADC-based state machine, especially since that means the ADC can't be
used until at least one completes the calibration step.  The existing
per-sensor approach also doesn't help with checking for temperature
changes.

Now we have API to read the die temperature, which is the temperature
that's most relevant to updating ADC calibration.

Provide a dedicated calibration state machine that initiates a
calibration when it's constructed (so it'll be ready in less than 1 ms),
checks for temperature changes, and initiates a recalibration when
necessary.  This can and should be used in parallel with any other
ADC-based sensors.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>