sensor/adc: enhance state machine to prevent resampling
A design goal for LPSM is that it's always safe to invoke lpsm_process()
even if there's nothing to do.  This goal was not satisfied by the ADC
lpsm_wrapper.

If there's only one ADCClient in the system then the queue operation
initiated in ENTRY_SAMPLE will almost certainly cause the machine to be
updated to EXIT_SAMPLE before lpsm_process() is invoked again.

If there are multiple clients, the queued operation may be blocked, and
a subsequent re-invocation of lpsm_process() would improperly attempt to
queue the operation again.

Add a new intermediate state MS_QUEUED where the machine will stay after
the operation is queued, until the queued operation is resolved.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>