periph/adc: add API to configure resolution and oversampling
Testing with mixed one-shot and scan ADCClient classes revealed that
CONFIG.BURST is not reliable: the transition from scan-mode to one-shot
mode corrupts the burst triggering for a subsequent scan-mode so that
samples are taken over all enabled channels, rather than dwelling on
each channel.

Rework the ADCClient API so that we can configure RESOLUTION and
OVERSAMPLING, and control whether the additional TASKS_SAMPLE triggers
required for non-zero oversampling are supplied through CONFIG.BURST or
an external process.

As a side effect the default behavior of all ADCClient instances is
reverted to the power-on-reset values:
* ADC uses 8-bit instead of 10-bit resolution unless overridden
* SAADC uses 10-bit resolution instead of 14-bit
* SAADC bypasses oversampling instead of using 16x oversampling
* SAADC disables CONFIG.BURST

As these defaults aren't very good instantiated clients should use the
resolution() API to set the desired behavior.  Note that the values of
the parameters to this function are series-specific.

See: https://devzone.nordicsemi.com/f/nordic-q-a/45339/saadc-burst-problems-in-scan-vs-non-scan-acquisitions
Signed-off-by: Peter A. Bigot <pab@pabigot.com>