passim: remove hard-coded assumptions that psel is on P0
Most of this code was written before nRF52840 support was added, and
would result in undefined behavior if pin selector for a P1 gpio was
used to index into the P0 PIN_CNF region.

Replace almost all hard-coded references to nrf5::GPIO with use of a
gpio::pin_reference instance.

A few references in example code or in board-specific code where P1 is
not relevant remain.  bme280 also remains because the psel is for CSn
which is optional, and gpio::pin_reference requires a valid pin.  bme280
needs to be reworked to support iface_config_type to provide the home
for the required gpio::generic_pin.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>