misc/lipomon: remove calibration support
The adc::voltage_divider used to measure battery power in this class is
not wrapped by sensor::adc::lpsm_wrapper because the lipomon
infrastructure uses its own LPSM to track Vchg and CSO signals.
Previously this LPSM replicated code similar to that in
adc::lpsm_wrapper to ensure the ADC was calibrated before sampling.
This code is not only redundant, it had at least one error in it
(returning the wrong error value on ADC failure).

Remove this capability.

In examples/misc/lipomon add the necessary calibrator instance, and
remove the redundant request for a startup battery voltage capture.  The
state machine automatically transitions to sample on startup.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>