passim: refactor flag values for invalid environmental readings
Replace all the literal 30000 instances with references to
sensor::INVALID_stdenv.  Provide API to test values for validity, and a
standard way of encoding information about the cause of invalid
readings.

This changes the documented invalid value for SDP8XX and SHT21 from
negative to positive but all existing processors use absolute value so
the impact is negligible.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>